Выполнено ДЗ из 5 раздела "Модуляция"
